+++ R/data.R
@@ -78,6 +78,6 @@
 #' from the origin.
 #'
 #' @source
-#' The data in `ineqmats` can be recreated with the command
+#' The data in `roth_ineqmats` can be recreated with the command
 #' `sapply(1:24, make_roth_ineqmat)`.
 "roth_ineqmats"